+++ src/pages/Facility/services/FacilityServices.tsx
@@ -1,13 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import { navigate } from "raviger";
 import { useTranslation } from "react-i18next";
 
-import ColoredIndicator from "@/CAREUI/display/ColoredIndicator";
-import CareIcon from "@/CAREUI/icons/CareIcon";
-import duoToneIcons from "@/CAREUI/icons/DuoTonePaths.json";
-
-import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
 import { EmptyState } from "@/components/ui/empty-state";
 
 import Page from "@/components/Common/Page";
@@ -16,68 +9,8 @@ import { CardListSkeleton } from "@/components/Common/SkeletonLoading";
 import useFilters from "@/hooks/useFilters";
 
 import query from "@/Utils/request/query";
-import { type HealthcareServiceReadSpec } from "@/types/healthcareService/healthcareService";
+import { ServiceCard } from "@/pages/Facility/settings/healthcareService/ServiceCard";
 import healthcareServiceApi from "@/types/healthcareService/healthcareServiceApi";
-
-type DuoToneIconName = keyof typeof duoToneIcons;
-
-function ServiceCard({
-  service,
-  facilityId,
-}: {
-  service: HealthcareServiceReadSpec;
-  facilityId: string;
-}) {
-  const { t } = useTranslation();
-  const getIconName = (name: string): DuoToneIconName =>
-    `d-${name}` as DuoToneIconName;
-
-  return (
-    <Card className="transition-all duration-200 hover:border-primary/50 hover:shadow-sm rounded-md">
-      <CardContent className="py-3 px-4">
-        <div className="flex flex-col gap-3 md:flex-row md:items-center">
-          <div className="flex items-center gap-3 flex-1 min-w-0">
-            <div className="relative size-10 rounded-sm flex p-4 items-center justify-center">
-              <ColoredIndicator
-                id={service.id}
-                className="absolute inset-0 rounded-sm opacity-20"
-              />
-              <CareIcon
-                icon={
-                  service.styling_metadata?.careIcon
-                    ? getIconName(service.styling_metadata.careIcon)
-                    : "d-health-worker"
-                }
-                className="size-6 relative z-1"
-              />
-            </div>
-            <div className="flex-1 min-w-0">
-              <h3 className="font-semibold truncate text-gray-900 text-base">
-                {service.name}
-              </h3>
-              <p className="mt-0.5 text-xs text-gray-500 truncate">
-                {service.extra_details}
-              </p>
-            </div>
-          </div>
-          <Button
-            onClick={() =>
-              navigate(
-                `/facility/${facilityId}/settings/healthcare_services/${service.id}`,
-              )
-            }
-            variant="outline"
-            size="sm"
-            className="px-3 text-xs whitespace-nowrap w-full md:w-auto md:ml-auto"
-          >
-            {t("view_details")}
-            <CareIcon icon="l-arrow-right" className="size-3" />
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
 
 export default function FacilityServicesPage({
   facilityId,
